requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-storage==2.13.0
requests==2.31.0
SQLAlchemy==2.0.23
alembic==1.13.1
email-validator==2.1.0.post1
Flask-Login==0.6.3
Werkzeug==3.0.1
pytest==7.4.4
black==24.1.1
flake8==7.0.0
flask-restx==1.3.0
Flask-Limiter==3.5.0
redis==5.0.1
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
Flask-Talisman==1.1.0
psutil==5.9.8
bleach==6.1.0
APScheduler==3.10.4

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.10
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.4